+++ processing_options.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"sync"
 
-	structdiff "github.com/imgproxy/imgproxy/struct-diff"
+	"github.com/imgproxy/imgproxy/structdiff"
 )
 
 type urlOption struct {
@@ -89,19 +89,27 @@ type gravityOptions struct {
 	X, Y float64
 }
 
+type extendOptions struct {
+	Enabled bool
+	Gravity gravityOptions
+}
+
 type cropOptions struct {
 	Width   int
 	Height  int
 	Gravity gravityOptions
 }
 
+type trimOptions struct {
+	Enabled   bool
+	Threshold float64
+}
+
 type watermarkOptions struct {
 	Enabled   bool
 	Opacity   float64
 	Replicate bool
-	Gravity   gravityType
-	OffsetX   int
-	OffsetY   int
+	Gravity   gravityOptions
 	Scale     float64
 }
 
@@ -112,8 +120,9 @@ type processingOptions struct {
 	Dpr          float64
 	Gravity      gravityOptions
 	Enlarge      bool
-	Extend       bool
+	Extend       extendOptions
 	Crop         cropOptions
+	Trim         trimOptions
 	Format       imageType
 	Quality      int
 	MaxBytes     int
@@ -140,8 +149,9 @@ const (
 	urlTokenPlain           = "plain"
 	maxClientHintDPR        = 8
 
-	msgForbidden  = "Forbidden"
-	msgInvalidURL = "Invalid URL"
+	msgForbidden     = "Forbidden"
+	msgInvalidURL    = "Invalid URL"
+	msgInvalidSource = "Invalid Source"
 )
 
 func (gt gravityType) String() string {
@@ -193,6 +203,8 @@ func newProcessingOptions() *processingOptions {
 			Height:       0,
 			Gravity:      gravityOptions{Type: gravityCenter},
 			Enlarge:      false,
+			Extend:       extendOptions{Enabled: false, Gravity: gravityOptions{Type: gravityCenter}},
+			Trim:         trimOptions{Enabled: false, Threshold: 10},
 			Quality:      conf.Quality,
 			MaxBytes:     0,
 			Format:       imageTypeUnknown,
@@ -200,7 +212,7 @@ func newProcessingOptions() *processingOptions {
 			Blur:         0,
 			Sharpen:      0,
 			Dpr:          1,
-			Watermark:    watermarkOptions{Opacity: 1, Replicate: false, Gravity: gravityCenter},
+			Watermark:    watermarkOptions{Opacity: 1, Replicate: false, Gravity: gravityOptions{Type: gravityCenter}},
 		}
 	})
 
@@ -415,17 +427,27 @@ func applyEnlargeOption(po *processingOptions, args []string) error {
 }
 
 func applyExtendOption(po *processingOptions, args []string) error {
-	if len(args) > 1 {
+	if len(args) > 4 {
 		return fmt.Errorf("Invalid extend arguments: %v", args)
 	}
 
-	po.Extend = parseBoolOption(args[0])
+	po.Extend.Enabled = parseBoolOption(args[0])
+
+	if len(args) > 1 {
+		if err := parseGravity(&po.Extend.Gravity, args[1:]); err != nil {
+			return err
+		}
+
+		if po.Extend.Gravity.Type == gravitySmart {
+			return errors.New("extend doesn't support smart gravity")
+		}
+	}
 
 	return nil
 }
 
 func applySizeOption(po *processingOptions, args []string) (err error) {
-	if len(args) > 4 {
+	if len(args) > 7 {
 		return fmt.Errorf("Invalid size arguments: %v", args)
 	}
 
@@ -447,8 +469,8 @@ func applySizeOption(po *processingOptions, args []string) (err error) {
 		}
 	}
 
-	if len(args) == 4 && len(args[3]) > 0 {
-		if err = applyExtendOption(po, args[3:4]); err != nil {
+	if len(args) >= 4 && len(args[3]) > 0 {
+		if err = applyExtendOption(po, args[3:]); err != nil {
 			return
 		}
 	}
@@ -471,7 +493,7 @@ func applyResizingTypeOption(po *processingOptions, args []string) error {
 }
 
 func applyResizeOption(po *processingOptions, args []string) error {
-	if len(args) > 5 {
+	if len(args) > 8 {
 		return fmt.Errorf("Invalid resize arguments: %v", args)
 	}
 
@@ -525,6 +547,21 @@ func applyCropOption(po *processingOptions, args []string) error {
 
 	if len(args) > 2 {
 		return parseGravity(&po.Crop.Gravity, args[2:])
+	}
+
+	return nil
+}
+
+func applyTrimOption(po *processingOptions, args []string) error {
+	if len(args) > 1 {
+		return fmt.Errorf("Invalid crop arguments: %v", args)
+	}
+
+	if t, err := strconv.ParseFloat(args[0], 64); err == nil && t >= 0 {
+		po.Trim.Enabled = true
+		po.Trim.Threshold = t
+	} else {
+		return fmt.Errorf("Invalid trim treshold: %s", args[0])
 	}
 
 	return nil
@@ -663,7 +700,7 @@ func applyWatermarkOption(po *processingOptions, args []string) error {
 		if args[1] == "re" {
 			po.Watermark.Replicate = true
 		} else if g, ok := gravityTypes[args[1]]; ok && g != gravityFocusPoint && g != gravitySmart {
-			po.Watermark.Gravity = g
+			po.Watermark.Gravity.Type = g
 		} else {
 			return fmt.Errorf("Invalid watermark position: %s", args[1])
 		}
@@ -671,7 +708,7 @@ func applyWatermarkOption(po *processingOptions, args []string) error {
 
 	if len(args) > 2 && len(args[2]) > 0 {
 		if x, err := strconv.Atoi(args[2]); err == nil {
-			po.Watermark.OffsetX = x
+			po.Watermark.Gravity.X = float64(x)
 		} else {
 			return fmt.Errorf("Invalid watermark X offset: %s", args[2])
 		}
@@ -679,7 +716,7 @@ func applyWatermarkOption(po *processingOptions, args []string) error {
 
 	if len(args) > 3 && len(args[3]) > 0 {
 		if y, err := strconv.Atoi(args[3]); err == nil {
-			po.Watermark.OffsetY = y
+			po.Watermark.Gravity.Y = float64(y)
 		} else {
 			return fmt.Errorf("Invalid watermark Y offset: %s", args[3])
 		}
@@ -758,6 +795,8 @@ func applyProcessingOption(po *processingOptions, name string, args []string) er
 		return applyGravityOption(po, args)
 	case "crop", "c":
 		return applyCropOption(po, args)
+	case "trim", "t":
+		return applyTrimOption(po, args)
 	case "quality", "q":
 		return applyQualityOption(po, args)
 	case "max_bytes", "mb":
@@ -789,6 +828,18 @@ func applyProcessingOptions(po *processingOptions, options urlOptions) error {
 	}
 
 	return nil
+}
+
+func isAllowedSource(imageURL string) bool {
+	if len(conf.AllowedSources) == 0 {
+		return true
+	}
+	for _, val := range conf.AllowedSources {
+		if strings.HasPrefix(imageURL, string(val)) {
+			return true
+		}
+	}
+	return false
 }
 
 func parseURLOptions(opts []string) (urlOptions, []string) {
@@ -984,6 +1035,10 @@ func parsePath(ctx context.Context, r *http.Request) (context.Context, error) {
 		return ctx, newError(404, err.Error(), msgInvalidURL)
 	}
 
+	if !isAllowedSource(imageURL) {
+		return ctx, newError(404, fmt.Sprintf("Invalid source"), msgInvalidSource)
+	}
+
 	ctx = context.WithValue(ctx, imageURLCtxKey, imageURL)
 	ctx = context.WithValue(ctx, processingOptionsCtxKey, po)
 
@@ -991,7 +1046,8 @@ func parsePath(ctx context.Context, r *http.Request) (context.Context, error) {
 }
 
 func getImageURL(ctx context.Context) string {
-	return ctx.Value(imageURLCtxKey).(string)
+	str, _ := ctx.Value(imageURLCtxKey).(string)
+	return str
 }
 
 func getProcessingOptions(ctx context.Context) *processingOptions {