+++ vips.c
@@ -338,7 +338,12 @@ vips_support_builtin_icc() {
 
 int
 vips_icc_import_go(VipsImage *in, VipsImage **out, char *profile) {
-  return vips_icc_import(in, out, "input_profile", profile, "embedded", TRUE, "pcs", VIPS_PCS_XYZ, NULL);
+  if (vips_icc_import(in, out, "input_profile", profile, "embedded", TRUE, "pcs", VIPS_PCS_XYZ, NULL))
+    return 1;
+
+  vips_image_remove(*out, VIPS_META_ICC_NAME);
+
+  return 0;
 }
 
 int
@@ -386,6 +391,48 @@ vips_flatten_go(VipsImage *in, VipsImage **out, double r, double g, double b) {
 
 int
 vips_extract_area_go(VipsImage *in, VipsImage **out, int left, int top, int width, int height) {
+  return vips_extract_area(in, out, left, top, width, height, NULL);
+}
+
+int
+vips_trim(VipsImage *in, VipsImage **out, double threshold) {
+  VipsImage *tmp;
+
+  if (vips_image_hasalpha(in)) {
+    if (vips_flatten(in, &tmp, NULL))
+      return 1;
+  } else {
+    if (vips_copy(in, &tmp, NULL))
+      return 1;
+  }
+
+  double *bg;
+  int bgn;
+
+  if (vips_getpoint(tmp, &bg, &bgn, 0, 0, NULL)) {
+    clear_image(&tmp);
+    return 1;
+  }
+
+  VipsArrayDouble *bga = vips_array_double_new(bg, bgn);
+
+  int left, top, width, height;
+
+  if (vips_find_trim(tmp, &left, &top, &width, &height, "background", bga, "threshold", threshold, NULL)) {
+    clear_image(&tmp);
+    vips_area_unref((VipsArea *)bga);
+    g_free(bg);
+    return 1;
+  }
+
+  clear_image(&tmp);
+  vips_area_unref((VipsArea *)bga);
+  g_free(bg);
+
+  if (width == 0 || height == 0) {
+    return vips_copy(in, out, NULL);
+  }
+
   return vips_extract_area(in, out, left, top, width, height, NULL);
 }
 
@@ -470,8 +517,8 @@ vips_arrayjoin_go(VipsImage **in, VipsImage **out, int n) {
 }
 
 int
-vips_jpegsave_go(VipsImage *in, void **buf, size_t *len, int quality, int interlace) {
-  return vips_jpegsave_buffer(in, buf, len, "profile", "none", "Q", quality, "strip", TRUE, "optimize_coding", TRUE, "interlace", interlace, NULL);
+vips_jpegsave_go(VipsImage *in, void **buf, size_t *len, int quality, int interlace, gboolean strip) {
+  return vips_jpegsave_buffer(in, buf, len, "profile", "none", "Q", quality, "strip", strip, "optimize_coding", TRUE, "interlace", interlace, NULL);
 }
 
 int
@@ -489,8 +536,8 @@ vips_pngsave_go(VipsImage *in, void **buf, size_t *len, int interlace, int quant
 }
 
 int
-vips_webpsave_go(VipsImage *in, void **buf, size_t *len, int quality) {
-  return vips_webpsave_buffer(in, buf, len, "Q", quality, "strip", TRUE, NULL);
+vips_webpsave_go(VipsImage *in, void **buf, size_t *len, int quality, gboolean strip) {
+  return vips_webpsave_buffer(in, buf, len, "Q", quality, "strip", strip, NULL);
 }
 
 int